Cumulative changes to support parsing NUL characters in prepared statements, without requiring binding.
Adds pragma allow_inline_nul which takes a simple boolean parameter =on/off to enable.
Adds sqlite3_sql_utf8() and sqlite3_expanded_sql_utf8() functions to be able to get a statement which has NUL characters in it, or has parameters with NUL characters in them bound to a parameter.
Adds test test/inlinenul.c to test basic functionality.
Has a slight improvement in speed for most tests.